Version 1.4
This could have been the long awaited version 1.0 - but instead I decided to synchronize the version numbering of this VIM plugin and the WOIM description itself. From now on the releases will be synchronized, with minor fixes in the VIM plugin released as minor releases (i.e. a fix to 1.4 would be 1.4.1 and would still be on par with the WOIM definition version 1.4). The version 1.4 of the WOIM definition adds time repetition as well as "checking" of todo-list items.

This release:

New feature: Added the option of checkboxes for items (Thanks to Christopher Truett (VIM script #3584). You can now easily add a checkbox in front of any item by <leader>v and subsequently toggle that checkbox via the same (<leader>v) or <leader>V if you want to add a time stamp to a box that you "check" (this also toggles the timestamp if you "uncheck" the item).

Fixes: Some minor cleanup.